% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Login.R
\name{getESRItoken}
\alias{getESRItoken}
\title{get ESRI arcgis online REST token}
\usage{
getESRItoken(username, secret)
}
\arguments{
\item{username}{string of username}

\item{secret}{string of password}
}
\description{
get ESRI arcgis online REST token
}
