% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandoF.R
\name{initCL}
\alias{initCL}
\title{Prepare for parallel processing Quickly}
\usage{
initCL(pcnt = 0.75)
}
\arguments{
\item{pcnt}{percent of cores to use}
}
\description{
Prepare for parallel processing Quickly
}
