% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownUtils.R
\name{knit_with_date}
\alias{knit_with_date}
\title{custom knit function}
\usage{
knit_with_date(input, output_dir = paste0(here::here(), "/Reports"), ...)
}
\arguments{
\item{...}{}
}
\description{
custom knit function
}
