% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotifyME.R
\name{notifyMe}
\alias{notifyMe}
\title{NotifyMe via email (use email version of phone for text)}
\usage{
notifyMe(
  sub,
  mes,
  whofrom,
  to = c("Michael <6099230973@vtext.com>"),
  username = "rmail@keepeverycog.com",
  secret,
  host = "mail.keepeverycog.com"
)
}
\arguments{
\item{sub}{subject}

\item{mes}{mesage text}

\item{whofrom}{message sent by - can use a function of process to help understand what part is done}

\item{to}{email address of recipient}

\item{username}{username for login to outgoing mail}

\item{secret}{password for login to outgoing mail}

\item{host}{host address for outgoing mail}
}
\value{

}
\description{
NotifyMe via email (use email version of phone for text)
}
