% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotifyME.R
\name{notifyMe}
\alias{notifyMe}
\title{Send an email (use phone number email to receieve a text)}
\usage{
notifyMe(
  message = "msg from R",
  mail_from = "rmail@keepeverycog.com",
  mail_rcpt = "6099230973@vtext.com",
  smtp_server = "mail.keepeverycog.com:465",
  use_ssl = "force",
  Username = "rmail@keepeverycog.com",
  Password = keyring::key_get("RMail_KEC", username = "rmail@keepeverycog.com"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Send an email (use phone number email to receieve a text)
}
